Update index.html
Ok, stabbing in the dark here (or, rather, in the .dart). "slideshow-text-1" and "slideshow-text-2" are now:
"Fraud and corruption *can* be stopped."
"It takes us all to be on board." 
All other "slideshow-text-X-Y" will be replaced by:
            <span class="upscaled" id="slideshow-text-3-1">Find.</span>
            <span class="upscaled" id="slideshow-text-3-2">Interpret.</span><br/>
            <span class="upscaled" id="slideshow-text-3-3">Resolve.</span> 
so that these 3 words should appear after "It takes us all to be on board," as below:
 "Find. Interpret.
     Resolve."
much in the same way slideshow text 4-1 to 4-4 appears now.
Changes to accompanying dart file: 
the eye slideshow should appear and fade out together with the new "slideshow-text-3-1,2,3".
I think the timing can remain the same as you have right now, I think it's pretty readable.
Hate giving up on trying to do all of this myself but other Hibis in house chores are piling up and I won't be able to keep up if I stubbornly stay on this. Nigel really wishes to send out emails to the other Hibis people today for them to check out the website...